Cost-Dashboard-Local/requirements.txt
azure-identity>=1.12.0
azure-mgmt-costmanagement>=3.0.0
azure-mgmt-resource>=22.0.0
azure-mgmt-containerservice>=20.0.0
matplotlib>=3.5.0
pandas>=1.3.0
python-dateutil>=2.8.2 
streamlit
reportlab
fpdf
forex-python